Fix Deletion of reservations and add price calculation

+++ src/main/java/de/cinetastisch/backend/dto/OrderResponseDto.java
@@ -1,7 +1,6 @@
 package de.cinetastisch.backend.dto;
 
 import de.cinetastisch.backend.enumeration.OrderStatus;
-import de.cinetastisch.backend.model.Reservation;
 
 import java.util.List;
 
@@ -9,7 +8,7 @@ public record OrderResponseDto(
         Long id,
         UserResponseDto user,
         OrderStatus orderStatus,
-        Integer total,
+        Double total,
         List<ReservationResponseDto> reservations,
         List<TicketResponseDto> tickets
 ) {
